I've been working into setting some things up to make this feel something you would like to draw on
- Tool Segregation:
Now all the tools used to edit a canvas are inherited a shared interface Tool, also made the brush is own thing rather than bein the default tool behaviour.

- Gui:

Decided to take a more functional approach to the Gui, right now is just a bunch of functions and some variables that define the Gui dimensions, functions are identified by "Gui" namespace

The only really new funtion is Gui::updatePanel, which resized the Gui to the current Screen size when called.

- Canvas class:
Added raylib Camera2D "canvasView", implemented, used in the _draw funtion to allow the user to move around and zoom in and out of the canvas.

* Next to do:

> Figure out how to calculate position fixed on the canvas (so you can actually draw to it...)

> Add Brushes to be selectable from Gui