Started with a dedicated module to manage mouse state, which subsequentially resulted in a series of changes on other parts of the program.
In order to check and interact with the mouse state `mState.hpp` must be added containing a series of functions that modify an mState struct which is only known by `mState.cpp`

This should in the future made easier to know whenever the mouse is already interacting with something and make the code more readable by encapsulating mouse state from the tools and gui logic.

mState:
- Added getter and setter of mouse position
- Added getter and setter of the mouse program state
- Added `Mouse::updateState()`, which must to be called on the main loop

Tools:
- Function `_drawToLayer(RenderTexture2D&, const Vector2&)` changed to `_drawTo(Canvas&)`
- Moved static variables in some tool clases to be variables of their respective class
- Module folder changed from `Brush` to `Tools`

Next I'm going to finish polishing `mState` and progresively integrate it in any place which uses mouse interaction.