Continuing Improving mouse module, trying to formet it in a more generalized way so I won't have to make a function for each parameter.
Now the `Mouse::State` object that the mouse state is written into can be obtained as a const reference so to read any element is needed. Depite this there's still some functions to directly get commonly used aspects of the mouse, such as its position and program state.

mState
- Added `Mouse::updateState()` function which updates the state of the mouse and must be called each frame.
- Added `Mouse::getMouseState()` to obtain a reference to the `Mouse::State` object used to save the mouse state.
- Ditched `Mouse::setPos(const Vector&)`, now the mState will fetch it position relative to the screen by itself on each `Mouse::updateState()` call.

Tools:
- Partly rewritten to accomodate to mState changes